% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_100norm_rel_abund.R
\name{compute_100norm_rel_abund}
\alias{compute_100norm_rel_abund}
\title{compute_100norm_rel_abund}
\usage{
compute_100norm_rel_abund(data)
}
\arguments{
\item{data}{a tibble containing the following column name: "rel_abund"}
}
\value{
data with a new column containing the normalized abundances
}
\description{
This function normalizes the relative abundances of the #' assigned formulas
to 100.
}
