% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_2sample_peak_differences.R
\name{compute_2sample_peak_differences}
\alias{compute_2sample_peak_differences}
\title{compute_2sample_peak_differences}
\usage{
compute_2sample_peak_differences(combined_samples)
}
\arguments{
\item{combined_samples}{a dataframe containing assigned formulas from
get_sample_data() for two different samples}
}
\value{
a dataframe containing the joined formulas and the differences in their
relative abundances
}
\description{
This function takes two dataframes containing assigned formulas, and returns
a dataframe containing the joined formulas, with relative abundance = 0 when
an assigned formula was not present in one of the samples.
}
\details{
The dataframe also returns a column containing the differential relative
abundance when sample 1 is subtracted from sample 2. Values > 0 have greater
relative abundances in sample 2, and values < 0 have greater relative
abundances in sample 1.

The function expects all of the static columns that are returned by the
get_sample_data() function, and will return an error if any of those column
names are missing from either sample. Specifically, those columns are:

"class_element", "class_hetero", "chem_formula", "theor_MI_mass", "theor_mz",
"DBE", "C", "H", "N", "O", "S", "DBEtoC", "ModAI", "HtoC", "OtoC", "NtoC",
"KMD_CH2", "z_CH2", "NOSC", "mass_defect"
}
