% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_arom_index.R
\name{compute_arom_index}
\alias{compute_arom_index}
\title{compute_arom_index}
\usage{
compute_arom_index(df, elements = c("C", "H", "N", "O", "S"), AItype = "ModAI")
}
\arguments{
\item{df}{a tibble containing a table of the assigned elements with a column name
for each element (e.g., "C", "H", "N", "O" and "S")}

\item{elements}{a character vector of the elements used for formula
assignment (default = CHNOS)}

\item{AItype}{a character string where "AI" = aromaticity index or "ModAI" =
modified aromaticity index}
}
\value{
a vector containing the aromaticity index for every assigned formula
}
\description{
This function is designed to compute the aromaticity index (AI) or modified
aromaticity index (ModAI) for the assigned molecular formulas in an MS
dataset. The calclulations are derived from Koch and Dittmar 2006
(http://dx.doi.org/10.1002/rcm.2386) with erratum
(http://dx.doi.org/10.1002/rcm.7433). The ModAI calculation treats one-half
of the oxygen as carboxyl oxygen, rather than carbonyl oxygen, in line with
published NMR data on marine DOM. This increases the number of compounds that
qualify as aromatic or condensed aromatic, but is still more conservative
than RDBE/C ratios. A (modified) aromaticity index >= 0.5 indicates an
aromatic structure, and a (modified) aromaticity index >= 0.67 indicates a
condensed aromatic structure.
}
