% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_AvgOSC.R
\name{compute_AvgOSC}
\alias{compute_AvgOSC}
\title{compute_AvgOSC}
\usage{
compute_AvgOSC(df, elements = c("C", "H", "N", "O", "S"))
}
\arguments{
\item{df}{a tibble containing MS data, including element numbers for each
elemental composition (e.g., C = 12, H = 26)}

\item{elements}{a character vector of the elements used for formula
assignment (default = CHNOS)}
}
\value{
a vector containing the AvgOSC for every assigned formula
}
\description{
This function calculates the average oxidation state of carbon (AvgOSC) for a
molecular formula, based on the formulas in Kroll et al. 2011, Riedel et al.
2012., and Lavonen et al. 2013. It is currently limited to CHNOS elements and
their associated oxidation states (H = 1, N = -3, O = -2, S = -2). P was
reserved due to its number of oxidation states (most common = -3, 3, and
especially 5). The halogens (X = -1) were also reserved.
}
