% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_DBEtoC.R
\name{compute_DBEtoC}
\alias{compute_DBEtoC}
\title{compute_DBEtoC}
\usage{
compute_DBEtoC(data)
}
\arguments{
\item{data}{a tibble containing the assigned molecular formulas}
}
\value{
data with a new column containing the computed DBE/C ratios
}
\description{
This function computes the DBE/C ratio, which is used as an estimate for
aromaticity.
}
