% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_elemental_ratios.R
\name{compute_elemental_ratios}
\alias{compute_elemental_ratios}
\title{compute_elemental_ratios}
\usage{
compute_elemental_ratios(data, num = "H", denom = "C")
}
\arguments{
\item{data}{a tibble containing column names for the desired elements (e.g.,
"H" or "C")}

\item{num}{character string for the element in the numerator (default = "H")}

\item{denom}{character string for the element in the denominator (default =
"C")}
}
\value{
data with a new column containing the computed elemental ratio
}
\description{
This function computes elemental ratios from user-specified elements.
}
