% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_KMD_CH2.R
\name{compute_KMD_CH2}
\alias{compute_KMD_CH2}
\title{compute_KMD_CH2}
\usage{
compute_KMD_CH2(df, selection = "theor", num_digits = 4)
}
\arguments{
\item{df}{a tibble containing MS data}

\item{selection}{a character string in c("meas", "theor") identifying the
desired calculation}

\item{num_digits}{an integer indicating the number of desired digits for the
KMD value (default = 4)}
}
\value{
df with a column containing the desired KMD values
}
\description{
This function computes the Kendrick mass defect (KMD) for a measured or
theoretical mass where CH2 (14.01565u) is the repeating unit (Kendrick 1963
and Hughey et al. 2001). The theoretical value is identical across members of
the same homologous CH2 series, and is useful when grouping for plots and
other purposes. The measured value is useful, especially in combination with
the appropriate z score, for identifying homologous CH2 series while
assigning formulas to measured m/z values. The number of digits can be
specified to limit rounding errors in the KMD calculation.
}
