% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_mass_defect.R
\name{compute_mass_defect}
\alias{compute_mass_defect}
\title{compute_mass_defect}
\usage{
compute_mass_defect(data)
}
\arguments{
\item{data}{a tibble containing the following column name: "theor_MI_mass"}
}
\value{
data with a new column containing the computed percent abundances
}
\description{
This function computes the mass defects of the assigned formulas.
}
