% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_nominal_mz.R
\name{compute_nominal_mz}
\alias{compute_nominal_mz}
\title{compute_nominal_mz}
\usage{
compute_nominal_mz(data)
}
\arguments{
\item{data}{a tibble containing the following column name: "mz"}
}
\value{
data with a new column containing the computed percent abundances
}
\description{
This function computes nominal mz values from a column containing mz values
(typically, either the assigned_formulas or the all_detected_ions dataframe.
}
