% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_NOSC.R
\name{compute_NOSC}
\alias{compute_NOSC}
\title{compute_NOSC}
\usage{
compute_NOSC(df, elements = c("C", "H", "N", "O", "S"))
}
\arguments{
\item{df}{a tibble containing a table of the assigned elements with a column name
for each element (e.g., "C", "H", "N", "O" and "S")}

\item{elements}{a character vector of the elements used for formula
assignment (default = CHNOS)}
}
\value{
a vector containing the NOSC for every assigned formula
}
\description{
This function calculates the average oxidation state of carbon (NOSC) for a
molecular formula, based on the formulas in Kroll et al. 2011, Riedel et al.
2012., and Lavonen et al. 2013. It is currently limited to CHNOS elements and
their associated oxidation states (H = 1, N = -3, O = -2, S = -2). P was
reserved due to its number of oxidation states (most common = -3, 3, and
especially 5). The halogens (X = -1) were also reserved.
}
