% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_perc_abund.R
\name{compute_perc_abund}
\alias{compute_perc_abund}
\title{compute_perc_abund}
\usage{
compute_perc_abund(data)
}
\arguments{
\item{data}{a tibble containing the following column name: "rel_abund"}
}
\value{
data with a new column containing the computed percent abundances
}
\description{
This function computes percent abundances from the relative abundances of
assigned formulas.
}
