% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_25perc_groups.R
\name{get_25perc_groups}
\alias{get_25perc_groups}
\title{get_25perc_groups}
\usage{
get_25perc_groups(data)
}
\arguments{
\item{data}{a tibble containing the assigned molecular formulas}
}
\value{
data with a new column containing the "25perc_group" to which each
  assigned formula belongs
}
\description{
This function groups assigned formulas into four groups ("top 25%", "second
25%", "third 25%", and "bottom 25%") by their percent contribution to total
assigned abundance.
}
