% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_detected_ions.R
\name{get_all_detected_ions}
\alias{get_all_detected_ions}
\title{get_all_detected_ions}
\usage{
get_all_detected_ions(no_hits, assigned_formulas)
}
\arguments{
\item{no_hits}{a tibble containing mz's and relative abundances for the "no
hits"}

\item{assigned_formulas}{a tibble containing the assigned formulas' mass
spectrometry data}
}
\value{
a tibble containing the combined mz's and relative abundances
}
\description{
This function combines the mz's and relative abundances of all detected ions
in the sample, including the detected ions that could not be assigned
formulas ("no hits").
}
