% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_elemental_ratios.R
\name{get_elemental_ratios}
\alias{get_elemental_ratios}
\title{get_elemental_ratios}
\usage{
get_elemental_ratios(sample_df, num = "H", denom = "C")
}
\arguments{
\item{sample_df}{a tibble containing the MS data for all assigned elemental compositions}

\item{num}{character string for the element in the numerator (default = "H")}

\item{denom}{character string for the element in the denominator (default = "C")}
}
\value{
sample_df with a column for the computed elemental ratio
}
\description{
get_elemental_ratios
}
\examples{
sample_df <- get_elemental_ratios(DOM_formulas_abbrev)
head(sample_df)
}
