% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_perc_abund.R
\name{get_perc_abund}
\alias{get_perc_abund}
\title{get_perc_abund}
\usage{
get_perc_abund(data)
}
\arguments{
\item{data}{a tibble containing the assigned molecular formulas}
}
\value{
data with a new column containing the computed percent abundances
}
\description{
This function computes percent abundances from the relative abundances of
assigned formulas.
}
