% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sample_data.R
\name{get_sample_data}
\alias{get_sample_data}
\title{get_sample_data}
\usage{
get_sample_data(
  file = file.choose(),
  ion_technique = "NegESI",
  elements_used = c("C", "H", "N", "O", "S")
)
}
\arguments{
\item{file}{a character string containing the filename and path for the Excel
file created by PetroOrg after molecular formula assignment}

\item{ion_technique}{a character string containing the MS ionization
technique used (default = NegESI)}

\item{elements_used}{a character vector containing the elements that were used
for molecular formula assignments (default = C, H, N, O, S)}
}
\value{
sample_data a named list containing 8 elements
}
\description{
This function reads sample Excel files exported by the MagLab's PetroOrg
software after molecular formula (i.e., elemental composition) assignment,
and returns a named list containing file and sample information, selected
mass spectrometry parameters, mass spectrometry data for the assigned
molecular formulas, mass spectrometry data for detected ions that could not
be assigned molecular formulas ("no hits"), and a table of all detected ions.
}
\details{
This function was built for MS data from Fourier transform ion cyclotron
resonance (FTICR) mass spectrometry using electrospray ionization in negative
ion mode (NegESI). This is the default ion type. Because natural organic
matter (NOM) is generally comprised primarily of the elements C, H, N, O, and
S, they comprise the default elements used for molecular formula assignment.

If the form of the files exported by PetroOrg changes, this function will
need to be revised.
}
