% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_VK_flat.R
\name{make_VK_flat}
\alias{make_VK_flat}
\title{make_VK_flat}
\usage{
make_VK_flat(data, plot_title = "")
}
\arguments{
\item{data}{a tibble containing the assigned molecular formulas}

\item{plot_title}{a character string containing the sample name}
}
\description{
This function makes a traditional van Krevelen diagram (H/c vs. O/C). This
plot shows the distribution of the assigned formulas as a function of
elemental ratios, but contains no information about the intensities of the
detected ions. The darker areas are where the highest number of formulas are
plotted.
}
\details{
The area above the horizontal line at H/C = 1.5 is designated as the
aliphatic region in accordance with D'Andrilli et al. 2015 and Lv et al.
2017. The area below the diagonal line from H/C = 1.1 is designated as the
aromatic region based on a modified aromaticity index >= 0.5 in accordance
with Koch & Dittmar 2006.
}
