% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hetero_class_multiple.R
\name{plot_hetero_class_multiple}
\alias{plot_hetero_class_multiple}
\title{plot_hetero_class_multiple}
\usage{
plot_hetero_class_multiple(data, plot_title = "")
}
\arguments{
\item{data}{a tibble containing containing the following column names: "class_hetero",
"perc_abund", and "sample_name"}

\item{plot_title}{a character string containing the sample name}
}
\description{
This function creates a heteroatom class bar plot for multiple samples. The total
number of heteroatom classes is typically very large, so it's useful to filter
the data to focus on the classes of interest first.
}
