% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hetero_class_single.R
\name{plot_hetero_class_single}
\alias{plot_hetero_class_single}
\title{plot_hetero_class_single}
\usage{
plot_hetero_class_single(data, plot_title = "", bar_color = "dark green")
}
\arguments{
\item{data}{a tibble containing containing the following column names: "class_hetero"
and "perc_abund"}

\item{plot_title}{a character string containing the sample name}

\item{bar_color}{a character string specifying the bar color}
}
\description{
This function creates a heteratom class bar chart for one sample. The total
number of heteroatom classes is typically very large, so it's useful to filter
the data to focus on the classes of interest first.
}
