% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Kroll_25perc_groups.R
\name{plot_Kroll_25perc_groups}
\alias{plot_Kroll_25perc_groups}
\title{plot_Kroll_25perc_groups}
\usage{
plot_Kroll_25perc_groups(
  data,
  plot_title = "",
  panel = FALSE,
  var_panel,
  num_col = 2
)
}
\arguments{
\item{data}{a tibble containing the following column names: "AvgOSC" (from
the computeAvgOSC() function), "C", "group_25perc" (from the
get_25perc_groups() function), and "rel_abund"}

\item{plot_title}{a character string containing the sample name}

\item{panel}{a logical value specifying whether a panel will be used (default
= FALSE)}

\item{var_panel}{a character string containing the column name specifying the
factor variable to be used for faceting}

\item{num_col}{an integer specifying the number of panel columns (default =
2)}
}
\description{
This function makes a Kroll diagram that plots average carbon oxidation
states against carbon number (Kroll et al. 2011). It also shows the
intensities of the detected ions by their percent contribution to total
assigned abundance.
}
