% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Kroll_differential.R
\name{plot_Kroll_differential}
\alias{plot_Kroll_differential}
\title{plot_Kroll_differential}
\usage{
plot_Kroll_differential(data, plot_title = "")
}
\arguments{
\item{data}{a tibble containing the following column names: "HtoC"
and "OtoC"}

\item{plot_title}{a character string containing the sample name (default =
none)}
}
\description{
This function makes a Kroll diagram that plots average carbon oxidation
states against carbon number (Kroll et al. 2011), but shows the intensities
of the peak abundance intensities between 2 samples gradient color scale
(as a function of relative abundance).
}
