% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_Kroll_flat.R
\name{plot_Kroll_flat}
\alias{plot_Kroll_flat}
\title{plot_Kroll_flat}
\usage{
plot_Kroll_flat(data, plot_title = "", panel = FALSE, var_panel, num_col = 2)
}
\arguments{
\item{data}{a tibble containing the assigned molecular formulas}

\item{plot_title}{a character string containing the sample name}

\item{panel}{a logical value specifying whether a panel will be used (default
= FALSE)}

\item{var_panel}{a character string containing the column name specifying the
factor variable to be used for faceting}

\item{num_col}{an integer specifying the number of panel columns (default =
2)}
}
\description{
This function makes a Kroll diagram that plots average carbon oxidation
states against carbon number (Kroll et al. 2011). It contains no information
about the intensities of the detected ions. The darker areas are where the
highest number of formulas are plotted.
}
