% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_VK_differential.R
\name{plot_VK_differential}
\alias{plot_VK_differential}
\title{plot_VK_differential}
\usage{
plot_VK_differential(data, plot_title = "")
}
\arguments{
\item{data}{a tibble containing the following column names: "HtoC"
and "OtoC"}

\item{plot_title}{a character string containing the sample name (default =
none)}
}
\description{
This function makes a traditional van Krevelen diagram (H/c vs. O/C), but
shows the intensities of the peak abundance intensities between 2 samples
gradient color scale (as a function of relative or abundance).
}
\details{
The area above the horizontal line at H/C = 1.5 is designated as the
aliphatic region in accordance with D'Andrilli et al. 2015 and Lv et al.
2017. The area below the diagonal line from H/C = 1.1 is designated as the
aromatic region based on a modified aromaticity index >= 0.5 in accordance
with Koch & Dittmar 2006.
}
