% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_VK_groups.R
\name{plot_VK_groups}
\alias{plot_VK_groups}
\title{plot_VK_groups}
\usage{
plot_VK_groups(data, var = "group_25perc", plot_title = "")
}
\arguments{
\item{data}{a tibble containing the assigned molecular formulas}

\item{var}{a character string containing the group factor column that will be
used to provide color (default = group_25perc)}

\item{plot_title}{a character string containing the sample name}
}
\description{
This function makes a traditional van Krevelen diagram (H/c vs. O/C). This
plot shows the distribution of the assigned formulas as a function of
elemental ratios, and shows the intensities of the detected ions by their
percent contribution to total assigned abundance.
}
\details{
The area above the horizontal line at H/C = 1.5 is designated as the
aliphatic region in accordance with D'Andrilli et al. 2015 and Lv et al.
2017. The area below the diagonal line from H/C = 1.1 is designated as the
aromatic region based on a modified aromaticity index >= 0.5 in accordance
with Koch & Dittmar 2006.
}
