% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorder_class_hetero.R
\name{reorder_class_hetero}
\alias{reorder_class_hetero}
\title{reorder_class_hetero}
\usage{
reorder_class_hetero(df)
}
\arguments{
\item{df}{a tibble containing a table of the assigned elements with a column name
for each element (e.g., "C", "H", "N", "O" and "S")}
}
\value{
a vector the heteroatom classes as ordered factors
}
\description{
This function orders heteroatom classes from O1 to Omax through CHNO_S1 to
CHNOS_max for use in summarizing and visualizing heteroatom class data. It
orders the factors based on a computed value, "level", which is discarded
after reordering. The levels increase in a nested fashion by number of O,
then number of N, and then number of S. The function is currently limited to
the default class of elements: C, H, N, O, S.
}
