% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_2sample_peak_differences.R
\name{compute_2sample_peak_differences}
\alias{compute_2sample_peak_differences}
\title{compute_2sample_peak_differences}
\usage{
compute_2sample_peak_differences(combined_samples, pos_1 = 1, pos_2 = 2)
}
\arguments{
\item{combined_samples}{a dataframe containing assigned formulas from
get_sample_data() for two different samples}

\item{pos_1}{an integer value indicating the order of sample 1 in the combined
sample dataframe}

\item{pos_2}{an integer value indicating the order of sample 2 in the combined
sample dataframe}
}
\value{
a dataframe containing the joined formulas and the differences in their
relative abundances
}
\description{
This function takes a dataframe containing the assigned formulas from at least
two NOM damples, and integer values representing the order of the two samples
that will be compared differentially, and returns a dataframe containing the
joined formulas, with relative abundance = 0 when an assigned formula is not
present in one of the samples.
}
\details{
The dataframe also returns a column containing the differential relative
abundance when sample 1 is subtracted from sample 2. Values > 0 have greater
relative abundances in sample 2, and values < 0 have greater relative
abundances in sample 1.

The function expects all of the static columns that are returned by the
get_sample_data() function, and will return an error if any of those column
names are missing from either sample. Specifically, those columns are:

"class_element", "class_hetero", "chem_formula", "theor_MI_mass", "theor_mz",
"DBE", "C", "H", "N", "O", "S", "DBEtoC", "ModAI", "HtoC", "OtoC", "NtoC",
"KMD_CH2", "z_CH2", "NOSC", "mass_defect"
}
