% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_CHNOS_element_class.R
\name{get_CHNOS_element_class}
\alias{get_CHNOS_element_class}
\title{get_CHNOS_element_class}
\usage{
get_CHNOS_element_class(data)
}
\arguments{
\item{data}{a tibble containing the assigned molecular formulas}
}
\value{
data with a new column containing the element classes as ordered
  factors
}
\description{
This function groups assigned formulas into element classes (e.g., CHO or
CHNOS) without regard to specific element numbers. It is intended to be used
for grouping in plotting, and has been written for the default element list
(CHNOS). The element classes are ordered by decreasing relative abundances,
so that the most abundant element class appears first.
}
