% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cum_perc_abund.R
\name{get_cum_perc_abund}
\alias{get_cum_perc_abund}
\title{get_cum_perc_abund}
\usage{
get_cum_perc_abund(data)
}
\arguments{
\item{data}{a tibble containing the assigned molecular formulas}
}
\value{
data with a new column containing the cumulative percent abundances
}
\description{
This function arranges percent abundances from least to greatest and computes
their cumulative percent abundances.
}
