% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_elemental_ratios.R
\name{get_elemental_ratios}
\alias{get_elemental_ratios}
\title{get_elemental_ratios}
\usage{
get_elemental_ratios(data, num = "H", denom = "C")
}
\arguments{
\item{data}{a tibble containing the MS data for all assigned elemental compositions}

\item{num}{character string for the element in the numerator (default = "H")}

\item{denom}{character string for the element in the denominator (default = "C")}
}
\value{
data with a new column containing the computed elemental ratio
}
\description{
This function computes elemental ratios from user-specified elements.
}
