% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sample_data.R
\name{get_sample_data}
\alias{get_sample_data}
\title{get_sample_data}
\usage{
get_sample_data(
  target_file = file.choose(),
  ion_technique = "NegESI",
  element_list = c("C", "H", "N", "O", "S")
)
}
\arguments{
\item{target_file}{a character string containing the filename and path for
the Excel file created by PetroOrg&copy after molecular formula assignment}

\item{ion_technique}{a character string containing the MS ionization
technique used (default = NegESI)}

\item{element_list}{a character vector containing the elements that were used
for molecular formula assignments (default = C, H, N, O, S)}
}
\value{
sample_data a named list containing file and sample information, MS
  parameter information, a tibble containing the MS data for all assigned
  elemental compositions, and a tibble containing the MS data for all
  detected ions that could not be assigned formulas ("no hits")
}
\description{
This function reads sample Excel files exported by the MagLab's PetroOrg
software after molecular formula (i.e., elemental composition) assignment,
and returns a named list containing file and sample information, selected
mass spectrometry parameters, mass spectrometry data for the assigned
molecular formulas, and mass spectrometry data for detected ions that could
not be assigned molecular formulas ("no hits").
}
\details{
This function was built for MS data from Fourier transform ion cyclotron
resonance (FTICR) mass spectrometry using electrospray ionization in negative
ion mode (NegESI). This is the default ion type. Because natural organic
matter (NOM) is generally comprised primarily of the elements C, H, N, O, and
S, they comprise the default elements used for molecular formula assignment.

If the form of the files exported by PetroOrg changes, this function will
need revisions.
}
