% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mass_spectrum.R
\name{plot_mass_spectrum}
\alias{plot_mass_spectrum}
\title{plot_mass_spectrum}
\usage{
plot_mass_spectrum(data, plot_title, max_x = NULL)
}
\arguments{
\item{data}{a tibble containing the assigned molecular formulas}

\item{plot_title}{a character string containing the sample name}

\item{max_x}{a numeric containing the maximum x value to plot (default =
NULL)}
}
\value{

}
\description{
This function plots the sample mass spectrum.The maximum x value can be
specified to produce figures for multiple samples on the same scale.
}
